Betere event-handeling voor zoeken zodat enter en klikken beter werkt.
