--- frontend/src/app/components/popover.css
@@ -1,13 +1,0 @@
-.popover-container {
-    position: relative;
-}
-
-.popover-content {
-    position: absolute;
-    top: 100%;
-    left: 100%;
-    transform: translateX(-100%);
-    margin-top: 8px;
-    background-color: var(--rvo-color-wit);
-    z-index: 1000;
-}
